moved nav bar items sealed class into skillet bottom nav bar file
