+++ app/src/main/java/com/kronos/skilletapp/SkilletNavigation.kt
@@ -59,41 +59,3 @@ class SkilletNavigationActions(private val navController: NavHostController) {
     }
   }
 }
-
-sealed class BottomNavItems<T : Route>(
-  val label: String,
-  val route: T,
-  val icon: ImageVector,
-  private val selectedIcon: ImageVector,
-) {
-  @Composable
-  fun Icon(isSelected: Boolean) {
-    AnimatedContent(
-      targetState = isSelected,
-      transitionSpec = {
-        fadeIn() togetherWith fadeOut()
-      }
-    ) {
-      if (it) {
-        Icon(imageVector = selectedIcon, contentDescription = null)
-      } else {
-        Icon(imageVector = icon, contentDescription = null)
-      }
-    }
-  }
-
-  data object RecipeList : BottomNavItems<Route.RecipeList>(
-    label = "Recipes",
-    route = Route.RecipeList(),
-    icon = Icons.AutoMirrored.Outlined.ListAlt,
-    selectedIcon = Icons.AutoMirrored.Filled.ListAlt
-  )
-
-  companion object {
-    val values by lazy {
-      listOf(
-        RecipeList,
-      )
-    }
-  }
-}
